fix: relocate `getFlangeBox` to the pipeblock class
