+++ src/main/java/supersymmetry/client/renderer/pipe/TanklessFluidPipeRenderer.java
@@ -13,7 +13,6 @@ import org.jetbrains.annotations.Nullable;
 
 import codechicken.lib.lighting.LightMatrix;
 import codechicken.lib.render.CCRenderState;
-import codechicken.lib.vec.Cuboid6;
 import codechicken.lib.vec.uv.IconTransformation;
 import dev.tianmi.sussypatches.api.core.mixin.extension.IconTypeExtension;
 import dev.tianmi.sussypatches.api.util.SusUtil;
@@ -32,6 +31,7 @@ import lombok.val;
 import supersymmetry.Supersymmetry;
 import supersymmetry.api.pipelike.CustomContext;
 import supersymmetry.api.unification.material.info.SuSyMaterialIconType;
+import supersymmetry.common.pipelike.tanklessfluid.BlockTanklessFluidPipe;
 import supersymmetry.common.pipelike.tanklessfluid.TanklessFluidPipeType;
 import supersymmetry.common.pipelike.tanklessfluid.tile.TileEntityTanklessFluidPipe;
 
@@ -104,32 +104,10 @@ public final class TanklessFluidPipeRenderer extends PipeRenderer implements Ico
     private void renderFlange(CCRenderState renderState, PipeRenderContext renderContext, EnumFacing side) {
         // bit 12 + side index = a cover sits on this side; recess the flange front face so it renders behind it
         boolean hasCover = (renderContext.getConnections() & 1 << (12 + side.getIndex())) > 0;
-        val cuboid = getFlangeBox(side, renderContext.getPipeThickness(), hasCover);
+        val cuboid = BlockTanklessFluidPipe.getFlangeBox(side, renderContext.getPipeThickness(), hasCover);
         for (val renderedSide : EnumFacing.VALUES) {
             renderOpenFace(renderState, renderContext, renderedSide, cuboid);
         }
-    }
-
-    public static Cuboid6 getFlangeBox(@Nullable EnumFacing side, float thickness, boolean hasCover) {
-        // Frame render box is offset by 0.001d inside, so here we offset by 0.0011d to make it rendered correctly
-        double min = Math.max((thickness > 0.3 ? 0.375d - thickness / 2.0d : 0.4375d - thickness / 2.0d), 0) + 0.0011d;
-        double max = 1 - min;
-        // the flange front face sits right at the base pipe cube's face when there is no cover, but is recessed
-        // (behind the cover box, which sits at 0.001 / 0.999) when a cover occupies this side
-        double faceMin = hasCover ? 0.0011d : -0.0001d;
-        double faceMax = 1 - faceMin;
-        double flangeMin = 0.1249d;
-        double flangeMax = 1 - flangeMin;
-
-        return switch (side) {
-            case WEST -> new Cuboid6(faceMin, min, min, flangeMin, max, max);
-            case EAST -> new Cuboid6(flangeMax, min, min, faceMax, max, max);
-            case NORTH -> new Cuboid6(min, min, faceMin, max, max, flangeMin);
-            case SOUTH -> new Cuboid6(min, min, flangeMax, max, max, faceMax);
-            case UP -> new Cuboid6(min, flangeMax, min, max, faceMax, max);
-            case DOWN -> new Cuboid6(min, faceMin, min, max, flangeMin, max);
-            case null -> new Cuboid6(min, min, min, max, max, max);
-        };
     }
 
     private static MaterialIconType getIconType(TanklessFluidPipeType pipeType) {
